feat(utils): Modularize formik validation objects
Add modules of formik initialValues, validationShema and form input labels  of BookingForm component, in the booking Form directory of utils directory.

This will make the code in the BookingForm component more concise. It will make it easier to manage the code and it makes the objects of formik validation more re-usable