Booking page feature (#5)
* feat(Booking): create booking page component

* style(Booking): Apply margin and padding

* feat(BookingForm): Create form component

* style(BookingForm): Apply style to form

* feat: Install formik and yup packages

* feat(BookingForm): use formik for validation

* syle(BookingForm): Apply styles

* feat(ErrorMessage): Add component

Add component that will display error messages while validating a form

* style(ErrorMessage): Apply styles

* feat(TextInput): Add new component

The TextInput component cotains the label, input area and error message.

This will make it easier to render other text input in the future.

* refactor(BookingForm): import and use components

Import and use ErrorMessage and TextInput Components.

* feat(toCamelCase): add new function

The toCamelCase conversts a string with white spaces, hyphens or underlines in to a camelCase string

* feat(DateInput): Add new component

* refactor(BookingForm): Import and use component

Import and use DateInput component

* feat(bookingHours): add new array of data

The data is an array of objects that contain an hour property and a status property .

This will be used as an internal data source for the select field (SelectInput component)

* feat(bookingOccasion): add new array of data

The data is an array of objects that contain an occasion.

This will be used as an internal data source for the select field (SelectInput component)

* feat(SelectInput): add new component

This component is used to render the select field and it's corresponding options

* chore(BookingForm.css): Update attribute selector

* refactor(BookingForm): Import component

Import the SelectInput component and substitute the select field of the form with the component.

The code becomes more concise and easier to manage

* feat(NumberInput): Add new component

* refactor(BookingForm): Import component

Import the NumberInput component and substitute the number input field of the form with the component.

* chore(utils): move js files to new directoriy

Move bookinHours.js and bookingOccasion.js to bookingForm directory

* feat(utils): Modularize formik validation objects

Add modules of formik initialValues, validationShema and form input labels  of BookingForm component, in the booking Form directory of utils directory.

This will make the code in the BookingForm component more concise. It will make it easier to manage the code and it makes the objects of formik validation more re-usable

* refactor(BookingForm): Add DRY code

* chore(validationData): Remove unnecessary code

* chore(SelectInput): Add key to option tag

* chore(Form): Rename BookingForm directory

Rename BookingForm directory to Form and provide BookingForm component with its own directory inside of the Form directory

* chore(BookingForm): Move component

Move component from "components/Form"  to "pages/Booking/SeconSection"

* feat(FirstSection): add new component

* refactor(Booking.jsx): reorganize code

* style(Booking): Apply styles

* chore(utils): Reorganize code

* feat(BookingForm): Add new props

These props will pass data from the child component to the parent component

* feat(utils): Add functionality

This functionality generates random available booking slots depending on the day that is set in the BookingForm component

This was supposed to be an API but the makers of the API served it with the wrong MIME type

* style(FirstSection): Apply styles

Apply styles to first section of Booking page

* refactor(SelectInput): change how prop is received

Receive the data prop in the SelectInput and OptionInput components as an array of strings

* feat(Booking): Pass callbacks as props

Pass functions as props to receive data from the BookingForm component

* feat(App): Add new route

This route will take  the user to ConfirmedBooking Page after a form is submited

* chore(bookingHours.js): remove unnecessary code

* feat(api.js): Add conditional statement

* styles(BookingForm): Apply styles

Increase the height of input areas

* feat(Button): Apply disabled functionality

The Button component receives a prop that should contain a Boolean value that will make the button become disabled or active

* feat(BookingForm): Apply conditional statement

The button will be set to disabled if the condition is true. The button is disabled or enabled  depending on the errors that are available while filling the booking form

* feat(ConfirmedBooking): Create page

This page is displayed after the booking form is submitted successfully

* feat(Booking): Add functionality  for navigation

Add functionality that will navigate to confirmed booking page after the form is submitted successfully

* style(App): Add new variable

* refactor(ConfirmedBooking): Change description

* style(ConfirmedBooking): Apply styles